refactor: refactor monitor notifications and incident tracking
- Refactor monitor creation and update logic to use a separate junction table for notifications.
- Introduce new tables for incidents and incident events.
- Add support for 'ping' monitor types.
- Change LatencyMs from int16 to int in notification messages.
- Update ping latency to use int instead of int16.
- Remove GroupID and NotificationIDs from the Monitor struct and related database columns.
- Implement CreateMonitorNotifications, DeleteMonitorNotifications, and GetNotificationIDsByMonitorID repository methods.
- Update mock repository with new methods for monitor notifications.
- Remove unused migration file for adding ping monitor type.

+++ core/notification/message.go
@@ -14,7 +14,7 @@ type MessageInput struct {
 	MonitorName string
 	Status      models.PingStatus
 	Region      string
-	LatencyMs   int16
+	LatencyMs   int
 	CheckedAt   time.Time
 	Detail      string
 }
