+++ models/ping.go
@@ -17,7 +17,7 @@ const (
 type Ping struct {
 	Time      time.Time       `json:"time" db:"time"`
 	MonitorID int64           `json:"monitor_id,string" db:"monitor_id"`
-	Latency   int16           `json:"latency" db:"latency"`
+	Latency   int             `json:"latency" db:"latency"`
 	Status    PingStatus      `json:"status" db:"status"`
 	Region    string          `json:"region" db:"region"`
 	Data      json.RawMessage `json:"data,omitempty" db:"data"`
